+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KXJ764GFGJJN0Y5F4E794JE1"
+        "01KXJVPSJVB6HNR3E13PM88TD3"
     )
     r"""The unique identifier of the data source"""