+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsehttp.py
@@ -55,9 +55,9 @@ class EvaluatorResponseHTTP(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-07-15T06:24:27.672Z"
+    created: Optional[str] = "2026-07-15T12:23:01.548Z"
 
-    updated: Optional[str] = "2026-07-15T06:24:27.672Z"
+    updated: Optional[str] = "2026-07-15T12:23:01.548Z"
 
     updated_by_id: OptionalNullable[str] = UNSET